refactor: change file name of provision to ss_provision
