+++ lib/fs_port.c
@@ -6,7 +6,7 @@
 #include <zephyr/storage/flash_map.h>
 
 #include "ss_cache.h"
-#include "provision.h"
+#include "ss_provision.h"
 #include "ss_profile.h"
 #include <onomondo/softsim/fs_port.h>
 #include <onomondo/softsim/list.h>
