+++ lib/ss_provision.h
